Move BuildDir to files
