+++ gradle-plugin-testing-junit/src/main/java/com/palantir/gradle/testing/files/BuildDir.java
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package com.palantir.gradle.testing.project;
+package com.palantir.gradle.testing.files;
 
 import com.google.errorprone.annotations.RestrictedApi;
 import com.palantir.gradle.testing.RestrictedCreation;
-import com.palantir.gradle.testing.files.FileFactory;
 import java.nio.file.Path;
 
 public record BuildDir(Path path) implements FileFactory {
